LLM_EVAL/requirements.txt
streamlit
pandas
requests
python-dotenv
openpyxl
matplotlib
pillow